rfxcom add support for serial over tcp
