+++ bundles/binding/org.openhab.binding.rfxcom/src/main/java/org/openhab/binding/rfxcom/internal/RFXComBinding.java
@@ -21,7 +21,7 @@ import javax.xml.bind.DatatypeConverter;
 import org.openhab.binding.rfxcom.RFXComBindingProvider;
 import org.openhab.binding.rfxcom.RFXComValueSelector;
 import org.openhab.binding.rfxcom.internal.connector.RFXComEventListener;
-import org.openhab.binding.rfxcom.internal.connector.RFXComSerialConnector;
+import org.openhab.binding.rfxcom.internal.connector.RFXComBaseConnector;
 import org.openhab.binding.rfxcom.internal.messages.RFXComBaseMessage.PacketType;
 import org.openhab.binding.rfxcom.internal.messages.RFXComMessageFactory;
 import org.openhab.binding.rfxcom.internal.messages.RFXComMessageInterface;
@@ -61,7 +61,7 @@ public class RFXComBinding extends AbstractBinding<RFXComBindingProvider> {
     @Override
     public void activate() {
         logger.debug("Activate");
-        RFXComSerialConnector connector = RFXComConnection.getCommunicator();
+        RFXComBaseConnector connector = RFXComConnection.getCommunicator();
         if (connector != null) {
             connector.addEventListener(eventLister);
         }
@@ -70,7 +70,7 @@ public class RFXComBinding extends AbstractBinding<RFXComBindingProvider> {
     @Override
     public void deactivate() {
         logger.debug("Deactivate");
-        RFXComSerialConnector connector = RFXComConnection.getCommunicator();
+        RFXComBaseConnector connector = RFXComConnection.getCommunicator();
         if (connector != null) {
             connector.removeEventListener(eventLister);
         }
@@ -151,7 +151,7 @@ public class RFXComBinding extends AbstractBinding<RFXComBindingProvider> {
         if (!provider.isInBinding(itemName)) {
             logger.debug("Received command (item='{}', state='{}', class='{}')",
                     new Object[] { itemName, command.toString(), command.getClass().toString() });
-            RFXComSerialConnector connector = RFXComConnection.getCommunicator();
+            RFXComBaseConnector connector = RFXComConnection.getCommunicator();
 
             if (connector == null) {
                 logger.warn("RFXCom controller is not initialized!");
@@ -172,7 +172,7 @@ public class RFXComBinding extends AbstractBinding<RFXComBindingProvider> {
     }
 
     private boolean executeCommand0(String itemName, Type command, final RFXComBindingProvider provider,
-            RFXComSerialConnector connector) {
+            RFXComBaseConnector connector) {
         String id = provider.getId(itemName);
         PacketType packetType = provider.getPacketType(itemName);
         Object subType = provider.getSubType(itemName);
